Added Docker file for CPU with gluonts preinstalled. Fixed most requirested changes. Updated documentation. Fixed bugs due to move.
