+++ src/gluonts/nursery/sagemaker_sdk/docker_images/with_gluonts_cpu/README.md
@@ -1,0 +1,5 @@
+0) You must have correctly set up your AWS CLI and must have Docker insatlled.
+1) Fill out the configurations <marked-like-this> in build_container_generic.sh 
+2) Copy requirements-extras-prophet.txt, requirements-extras-r.txt and 
+    requirements.txt from "gluon-ts/requirements" in this folder
+3) Execute build_container.sh with "./build_container.sh" in the terminal
